backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
pydantic
python-dotenv
httpx
openai